modules/helper/subtitles/requirements.txt
paho-mqtt>=1.4.0
srt